pyproject.toml
[project]
name = "movie-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-generativeai>=0.8.5",
    "groq>=0.31.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "rapidfuzz>=3.14.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "transformers>=4.56.1",
]
